FourProblemSetsInOne-UsingNodes
Problem 0 – Find Minnie

Implement the method getMin, which takes a single Node reference, head, that points to the beginning of a linked list. getMin should return the Node whose data is the minimum of the list. If head is null, then there is no minimum and the method should return null.

Problem 1 – Merging Two Linked Lists

Implement the method, mergeTwoLists, which merges two sorted linked lists and returns it as a single sorted list. The new list should be made by splicing together the nodes of the first two lists. Return the head of the merged linked list.

Problem 2 – tsiL dekniL a esreveR

Write the method, reverseList, which reverses a linked list. The original head should be the last element of the new list, the second node should become the second to last node, etc. This method returns the head of the newly reversed list.

Problem 3 – Was it a rat I saw?

A palindrome is read the same forwards as it is backwards. Implement the method, isPalindrome, which determines if a linked list is a palindrome.